Fix reversed arguments to eig() call in chebT1rtsmatgep() in chebfun2v.roots().
These should be the other way around:  the leading-order coefficient
should be the "B" argument, while the constant term should stand in for
"A".  This closes #1505 and is the right resolution for #1232.
chebfun2v.roots() now handles bivariate linear problems correctly.

This commit also reverts the changes made in 8bcf6c4d and 6bb4be81 to
address #1232, as the cutoff tolerance is not the issue here.  The
reason that making that tolerance smaller appeared to fix things is that
it sometimes caused a small leading-order coefficient that in principle
_should_ be stripped to not be stripped, so that when the coefficients
were passed to chebT1rtsmatgep() later, the input would have size at
least 3 in the third dimension.  This caused the handler for the general
case, which does things correctly, to get called instead of the handler
for the linear case, which was faulty prior to the changes in this
commit.

A test has been added to hopefully prevent this behavior from recurring
in the future.