Fix delta colors when plotting delta-quasimatrices on R2014b.
This was subtler than I thought it would be.  If you want the colors to
be right, you can't just throw all delta data from all columns into a
struct and work with that, since you've discarding the information about
which column they came from, and that's important to determining color.

To address this, we turn deltaData into a cell array of cell arrays of
data for plotting delta functions, where each inner cell array contains
all the data for plotting the deltas for one column.  The contents of
the inner cell arrays are precisely the arguments which need to be
passed to the new plotDeltasForOneFunction() subroutine, which is
actually just the old plotDeltas() subroutine, renamed and tweaked just
a little bit.  The new plotDeltas() just calls
plotDeltasForOneFunction() for each column after ensuring the colors get
set correctly by using the ColorOrderIndex graphics property.

This fix applies to R2014b and later only, since ColorOrderIndex is not
available on earlier versions of MATLAB.

At some point, we really need to redesign the plotting stack from the
ground up.  The data structures are not clearly documented, and the code
in here is convoluted and hard to follow in places.  This fix, being the
ugly hack that it is, makes this problem worse, but I don't know a
better solution at the moment.