Restore enableFunqui preference.
I'm about to make some changes which officially require that the "tech"
field of a chebfunpref object be a handle to a tech constructor (or
something equivalent).  This means that we can no longer handle funqui
by storing the string "funqui" in the "tech" preference field and
parsing it out later in smoothfun.

Frankly, I never liked this approach to handling funqui in the first
place, precisely since it means we can't make the above assumption.
Very early on when working on preferences, we dealt with funqui by
having an upper-layer preference dedicated to it (see bd9081ed);
however, it was quickly removed in favor of the present system (see
82fa69d4), since it was judged that funqui should be kept hidden from
users (except via the "equi" flag) on the basis that Chebfun is not a
general-purpose system for interpolation.

This commit is a compromise:  it restores the enableFunqui preference
but does not print it in display() or list it in the documentation, so
from the average user's perspective, it does not exist.  As an
additional benefit (albeit a minor one), things like

  f = chebfun(equi_data, 'equi', 'tech', @chebtech1);

now work as expected.  Prior to this commit, for such a call, the input
parser, upon encountering the "equi" flag, would set the "tech"
preference to "funqui", but this would then get overridden by next pair
of inputs, which set "tech" to @chebtech1.  (To get around this, instead
of setting "tech", you could set the undocumented preference
"funquiTech", which this commit does away with.)