Get all default tech preferences in chebfunpref from tech.techPrefs.
In particular, this applies for the "abstract" tech preferences that the
upper layers are allowed to use.  (The other tech preference defaults
were always retrieved from the techs when needed.)  Previously, we
stored separate defaults for these inside chebfunpref even when the user
did not override them, but this means that techs can't supply defaults
appropriate for themselves.  For an example of the consequences of this,
see GitHub issues #1336 and #1337.

I feel pretty strongly that this is the "right" thing to do.  The only
reason we did the other stuff before was for convenience and to avoid
the performance hit associated with this.  We'll see if the latter is
enough of an issue to worry about.  At the very least, it seems to have
a negligible effect on the timing of chebtest('chebfun'), at least on my
machine.

I've also gotten rid of the documentation in @chebfunpref/chebfunpref.m
that lists the default values for the "abstract" preferences.  These
defaults are no longer accurate, since they're now determined by the
techs.  To get the default values, one can consult the documentation for
the relevant techPref() function or just call chebfunpref() without any
arguments and see what they are.