Primitive implementation of "turbo" chebtechs.
This is almost exactly the code from my January mini-project.  The only
difference is that it uses LNT's choice for the ellipse of integration
(2/3 power of the Chebfun ellipse; I used the Chebfun ellipse itself)
and number of quadrature nodes (4 times the number of coefficients
desired; I used 5 times).  Usage is controlled by a new "useTurbo"
chebtech preference.