Added a constructor that applies complete pivoting sperately to the two matrices (functions) on the right of the equation: % [B    C]  = 0.5*[ (B+C)   (B+C)]  + 0.5*[  (B-C)    -(B-C)] % [JC  JB]        [J(B+C)  J(B+C)]        [-J(B-C)    J(B-C)]
This code does not call the chebfun2 constructor on each piece, but implements the complete pivoting strategy.  It exploits symmetries during the construction that chebfun2 would not.

You can use this version of the constructor as:

>>g = spherefun(f,'PlusMinus');

Hopefully we can show that this way of doing things has some fundamental flaw...