Set tech preference in chebfun.compose().
After adding the "useTurbo" preference, the following code began issuing
warnings, since the chebtech and trigtech preferences now no longer
overlap:

    z = chebfun('4*exp(1i*s)', [0 2*pi], 'trig');
    f = exp(z);

The reason is that in chebfun.compose(), we create a chebfunpref() with
the default preferences, which has the tech set to chebtech2.  The
preference object's techPrefs are therefore those of chebtech2, but the
compose() method that ultimately gets called at the bottom of the stack
in the above is the one in trigtech, so we ended up passing chebtech
preferences to trigtech.

We fix this by setting the "tech" preference correctly in
chebfun.compose().  I don't know if this is the best solution, and in
particular, I'm not sure what is the right thing to do when composing
two chebfuns of different techs.  At the moment, the way the code is
written dictates that compose(f, g) will call the compose methods for f
instead of those of g, and the fix goes along with this.