Remove epslevel field from chebtech.
epslevels were effectively disabled with #1474, but 99% of the epslevel
adjusting code was left in place.  That code has now been effectively dead for
months, and we haven't run into any problems.  This commit is the first step
towards finishing the process of getting rid of epslevels entirely.  Doing this
will greatly simplify many parts of the code and make things substantially less
confusing.

Previously, we decided to leave the epslevel code dormant with the idea that it
might one day be resurrected if one of us comes up with a good epslevel
concept.  It seems fair to say that none of us have really expressed any desire
to do this, nor have any of us come forth with good ideas for how to do so.  It
also seems very unlikely that this situation will change in the short (or even
medium) term.  Moreover, there are good arguments against the epslevel concept
that have been established that have broad buy-in among the team, and even if
epslevels do end up being revived, it is quite likely that the existing
framework would be inappropriate and would have to be scrapped anyway.

For these reasons, we have judged that the benefits of leaving the old code in
place are outweighed by the disadvantages of keeping it around and therefore
have decided to complete the removal process.

Note that we have created a chebtech.epslevel() method and continue to support
get(f, 'epslevel') for chebtechs at the moment.  These are temporary measures
to keep enough of the system working that we can remove epslevel in stages
instead of all in one go.  They will be removed in a later commit.