Remove usage of epslevel within chebfun methods.
Note that we haven't removed chebfun.epslevel() and the clause in chebfun.get()
for retrieving a chebfun's epslevel, as these are still used externally many of
the tests.

We also removed support for get(f, 'hscale-local') (where f is a chebfun), as
it seems this was not being used at all by anyone.