Adjust simplify tolerances using local and global vscales.
We removed these lines when removing epslevel from chebfun.simplify(), but
after discussing things with Jared Aurentz, it seems that they may still have a
place for the time being.  (For instance, if a chebfun has one fun with a
vscale of 1 and one with a vscale of 1e50, then the vscale-1 fun should
probably simplify to zero, since it is negligibly small compared to the overall
vscale of the chebfun.)

Nevertheless, it is worth noting that removing these lines caused none of the
tests to break.