Force get() call in chebfun.plotregion() to return outputs in a cell array.
The later code expects a cell input.  Without this, we get a "cell
contents reference from a non-cell array object" error when calling
plotregion() on a chebfun with one fun (the most common case).  The
reason is that get(h, 'Color') will not pack its output in a cell array
when h is only one line object.  Placing the string 'Color' in a cell
array guarantees that the output will always be packed in a cell.