Set default MarkerSize to 20 in separableApprox.waterfall().
With a default LineWidth of 2, a default MarkerSize of 7 makes the
markers too small to be visible on my machine.

A change of this sort was originally effected in 33854a3e, but this
change was probably undone when separableApprox got merged into
development and replaced chebfun2.waterfall() with
separableApprox.waterfall().