Overhaul chebfun.prod().
This function had several problems (not that anybody cared, as this
function is basically never used):

 - It was returning a chebfun of length 1 instead of scalar output.
   (See #1436.)

 - It was not properly respecting transposition states (fixed with the
   previous commit and this one).

 - It did not have a dimension argument, making it inconsistent with
   MATLAB prod(), which lets you take products along discrete
   dimensions.

This commit fixes all of these issues and adds a test for this function.

cumprod() probably needs a similar overhaul, but as that function's used
even less frequently than this one, I'm not going to deal with it right
now.