Build final chebfun2 from an empty chebfun2 in chebfun2.outerProduct().
Building it from a zero chebfun2 causes the pivotLocations information
to be set, but that's nonsensical here, since we aren't constructing the
chebfun2 using the adaptive Gaussian elimination procedure.

This closes #1331.  The problem in that ticket arose because
chebfun2.chol() was using the bogus pivot information returned by
chebfun2.outerProduct() to do something.  Added a test for this example
to ensure it doesn't recur in the future.