Completely remove local transform functions from ratinterp().
This completes the work begun in the previous commit, closing #1381.  It
turns out that the modifications to ratinterp() to completely remove the
local DCT/IDCT implementations were not too difficult.  In addition to
removing another 30 or so lines from ratinterp(), these changes make the
code easier to understand, as it is now obvious exactly how the
values-to-coefficients and coefficients-to-values maps come into play.

A few notes:

 - The only reason we needed the odd diagonal scaling in the local
   ratinterp() functions dct1() and idct1() was due to the peculiar way the D
   and Z matrices were being constructed for first-kind Chebyshev
   points in assembleMatrices().  By changing the construction method to mimic
   that already used for second-kind Chebyshev points, we were able to get rid
   of the diagonal scaling altogether, along with some mysterious transposes.

 - After the above change, dct1(), idct1(), dct2(), and idct2()
   just became direct wrappers for the corresponding vals2coeffs() and
   coeffs2vals() functions in chebtech[12], which was what allowed us to
   eliminate them entirely.

 - There seems to be some bizarre MATLAB parser bug that prevents us
   from passing the function handle @chebtech1 as the 'tech' parameter to the
   chebfun constructor calls in constructRatApproxCheb1() after we
   invoke chebtech1.coeffs2vals() earlier in the function.  To get
   around this, we store the handle to the chebtech1 constructor in a variable
   at the start of the function and use that instead.

For reference, the following simple example will reproduce the parse
error mentioned above.  Place the following function in an M-file
temp.m in a folder on the MATLAB search path:

function temp()
        v = chebtech1.coeffs2vals([5 ; 4 ; 3 ; 2 ; 1])
        f = chebfun(v, 'tech', @chebtech1)
        v = chebtech1.coeffs2vals([5 ; 4 ; 3 ; 2 ; 1])
        f = chebfun(v, 'tech', @chebtech1)
end

Calling temp() then produces the following error:

"chebtech1" was previously used as the prefix of a compound name,
conflicting with its use here as the name of a function.

Getting rid of the last line in the function body makes the error go
away.  This behavior is reproducible for me in R2013b and R2014b,
running on Linux.