Ensure d <= n in funqui exceptional case.
If the number of points used by funqui is very small, say, 2 or 3, this case
will be hit.  Without this change, it will set d = 4, which is bad, since the
FH scheme requires d <= n.  Messing this up causes the fhBaryWts routine to
return 0 for all the weights, which results in NaNs when the function handle is
sampled.  So, we use the smaller of d = n and d = 4.

Miraculously, this doesn't cause any crashes with the way things are written at
the moment.  This is a byproduct of the fact that we extrapolate out NaNs
during the construction process---even when there are _lots_ of them---and the
fact that the way the check in @chebtech/sampleTest.m is written happens to
pass for NaN inputs due to the fact that comparing anything with a NaN returns
false.  In fact, the way this issue was discovered was that Jared Aurentz
stumbled across the problem when trying to modify chebtech.sampleTest().

Aside from the main issue, JA also points out that the behavior of
chebtech.populate() is a little concerning here, as it will happily extrapolate
out as many NaNs/Infs as it sees, even if those greatly outnumber the number of
legitimate samples it picked up.  I've added a TODO to @chebtech/populate.m
about this.

Unfortunately, there is currently no way to test for this behavior, since
funqui is totally internal to smoothfun.  In fact, the second test in
tests/chebfun/test_constructor_equi.m excites the problem, but the system does
not crash for the aforementioned reasons.  To give us at least a little
testability here, I'll change the way the check is done in sampleTest()
shortly.