Call happinessCheck() in chebfun.pdeSolve() instead of specific check.
The specific check that gets used should be the default one (which, at
the moment, is standardCheck()).  If the default needs to be overridden,
that should be done via preferences passed to happinessCheck(), not by
invoking the specific check function directly.

If, for some reason, it is absolutely necessary to invoke the specific
check directly, one must provide a complete set of arguments, as
happinessCheck() does the common input parsing for the specific checks.
The line changed in this commit was invoking standardCheck() with an
empty data argument, and this caused a few of the chebop2 tests to
crash, along with the test for pde15s().