Fix dimension mismatch bug in plotcoeffs(f, 'loglog') for trigtech.
This bug appeared as a result of the changes to trigtech.plotcoeffs()
made in 7b96235a (Remove epslevel from trigtech.), though it has
actually existed in trigtech for longer.  The reason we never came
across it before is because the branch of code it was on could not be
reached by calling plotcoeffs() on a trigtech-based chebfun.

Prior to 7b96235a, trigtech.plotcoeffs() had an internal boolean
variable named plotEpsLevel that controlled whether the epslevel would
be plotted along with the coefficients.  plotEpsLevel variable was set
to true by default and could be overridden by a 'noepslevel' flag
argument passed in by the upper layers.  chebfun.plotcoeffs(), however,
never passed this flag down even when it was given 'noepslevel' itself,
so the parts of the code that were always executed were the branches
corresponding to plotEpsLevel being true.  But when getting rid of
epslevel in 7b96235a, we naturally kept the branches of the trigtech
plotting code corresponding to the plotEpsLevel variable being false.

This commit fixes this error by using the plotting call from the
plotEpsLevel = true code branch, which was correct.  (Indeed, I don't
understand why the call wasn't just factored out of the two branches in
the first place.)  Tests for plotcoeffs() involving trigtech have been
added to tests/chebfun/test_plotcoeffs.m.