Pass dimension to cumsum() in computeDerCoeffs() in chebtech.diff().
This is necessary because if the degree is low (2, 3), we'll end up
calling cumsum() on a row vector, and when operating on a vector MATLAB
cumsum() goes along the non-singleton dimension.  That's not desirable
in this case:  we always want to cumsum down _columns_ here, even if the
vector is a row.

A test has been added to check for this issue.