Set tolerance for chebfun2 outer product tests in tests/chebfun/test_mtimes.m.
The recent adjustments to simplify seem to have caused one or two of
these tests to just barely fail.  I was going to bump the tolerances for
the failing tests, but upon closer inspection, I found that they were
using a tolerance set for a previous test in the file.  This is almost
certainly a mistake.

To fix this, I've copied over the tolerance from the test for kron() (as
these tests mimic the ones in that file) and adjusted it by a factor of
10 to account for the inline adjustments made to it in this test.