Fix string concatenation bug in chebfun2.explain().
Prior to this commit, calling explain(f) with

  f = chebfun2(@(x, y) sin(x + y))

fails.  The reason is that when we get to the changed line, the variable
str is already a cell array, so when we do str = {str ; '...'}, we wind
up with a two-element cell array, the first element of which is a cell
array and the second of which is a string.  This is not what we want: we
really just want to append a new cell containing the string  '...' to
the end of str.

This suggests that we change the line to str = [str ; '...'] (or,
apparently equivalently, str = [str ; {'...'}]), but if we do this, then
the function crashes when we try it with

  f = chebfun2(@(x, y) sin(x + y) + x.*y)

instead because for this function, str is a string when we hit the
changed line, not a cell array.

The right thing to do here would be to rewrite explain() so that str is
_always_ a cell array, even when it could legally be a string, but
that's not worth the effort at the moment.  Instead, we just turn it
into a string here with the desired text.  This will clear any other
text from the text box when it gets displayed, but I would argue that
this is a natural thing to do at this point in the presentation anyway.