Explicitly check for the presence of imfreehand() in +cheb/choreo.m, etc.
I think this makes the intent clearer, plus using try-catch block in the
way done has some unexpected effects.  If one of the later lines in the
try-block fails (as may happen if one inputs a 1-point "curve"), instead
of a crash or a proper recovery from the error, the user will be
re-prompted to input a curve, this time using ginput() instead.