Detect exps. when those of addends differ by non-integers in singfun.plus().
The heuristic being used here doesn't have any real justification behind
it and, in fact, gives the wrong answer in many cases.  For instance, if
you're adding 0 to sqrt(x) (assuming that we weren't trapping this as a
special case) on [0, 1], the heuristic gives that the exponents of the
answer are [0 0], while they should be [0.5 0].

So, we revert to just auto-detecting them.  This in itself is mostly
meaningless, as the answer isn't actually representable using singfun,
but this way, we're at least making a "good faith effort" to determine
the "exponent" instead of just guessing.

It might actually be better just to issue an error here and forbid these
additions entirely, but that's a discussion for another day.