Revert "Detect exps. when those of addends differ by non-integers in ..."
This reverts commit 33aebc678f1af6e6ac24e91afc26e101f747cf3f.

This change caused test #12 in tests/chebfun/test_cumsum.m to fail.
Apparently, in that particular case, the heuristic gets the right
exponent, and the auto-detection procedure fails.  As this isn't a big
issue, and since I don't have a good idea for fixing the auto-detection
here, I'm just going to revert the change for the time being.