Pass vscale to smoothfun constructor in singfun.plus().
This partially reverts f337559b.  One of the changes in that commit
removed the lines that pass the vscale information down to the smoothfun
constructor when we hit the case in which the exponents differ by
non-integer values in singfun.plus().  This was causing the example
cheb/Convergence.m to hang because one of the functions it constructs (a
rare instance of a case that actually hits this code and an even rarer
example of one where the construction succeeds) is built by subtracting
two functions that are very close, and the constructor was failing due
to the resulting noise.