Replace mod() with rem() in check in even-length trigBary().
If s is small and negative (as can happen when s should be exactly 0 in
theory but rounding error in the sum causes things to happen otherwise),
mod(s, pi) is close to pi.  That's not what we want here:  we really
want rem(s, pi), which will yield a number near zero.

Closes #1743.