Properly check for equispacedness in trigBaryWeights().
Instead of just checking that the points are equispaced, we check that
they are equispaced _with the proper spacing_.  This prevents things
like the grid [-1 -1/3 1/3 1] from being seen as equispaced on [-pi,
pi].

Also clarified the trigBaryWeights() documentation to indicate the fact
that it computes weights for 2*pi-periodic interpolation.  That it
behaves like this has always been true, but it hasn't been written down
anywhere.

Closes #1744.