Significant overhaul to how QR is performed for CHEBFUN objects with multiple pieces.
In particular (where possible) the QR factorization is performed individually on each piece (at the lower levels) and the results the re-combined at the CHEBFUN level. (Essentially "panel-factored QR".)

Performance needs to be investigated, but this has removed a hundred or so lines of code from chebfun/qr (in addition to removing a horrible check for whether the funs consisted of 'simple' chebtech objects.