Add chebfun.tolUnion() and chebfun.tolUnique().
These methods act just like built-in union() and unique() except that
they use a tolerance to decide when two floating-point numbers are
equal.  If the combined vector has several nearly-equal values, the
behavior may not be "correct" (see the TODO in these files), but this
not likely to be an issue in practice.