Remove unnecessary method declarations from classdef files.
These just waste space and are a recipe for confusion because MATLAB
doesn't require them to be kept in-sync with the declarations in the
M-files that actually implement the methods.  This can cause bizarre
things to occur.

Say, for instance, that you are working on a method which is delcared as
having 2 arguments in the classdef file, but, as part of your work, you
add a third in the M-file.  If you forget to update the declaration in
the classdef file, then when you call the method with 3 arguments,
MATLAB will silently take the classdef-declared prototype as the "right"
one and give you a "too many arguments" error. This looks really
mysterious if you've forgotten that the classdef declaration even
exists.

Closes #1601 and #888.