Ensure minimum grid lengths in the chebfun2 constructor have the correct form.
For better or worse, the chebfun2 constructor makes assumptions that the
grid for constructions based on trigtech is a power of 2 and that the
grid for constructions based on chebtech is one plus a power of 2.  This
code ensures that this is the case.

This commit is just a fix-up of Alex Townsend's commit from #1771 to
work in the world of #1765.