Revert "modified trigtech and chebtech innerProducts to handle mixed tech cases"
This reverts commit 1a54bfe93579243e4e09aff6c747e1acea24e43a.

These checks really should not be done at the tech level because it
doesn't scale and because it violates our long-standing principle that
the techs should be detachable from the rest of the system.  This should
be handled inside the chebfun class, where it is for all the other
situations in which we have to do such conversions:  we don't implement
any other binary operations between objects of mixed types at the tech
level.