added about 200 more test cases to chebfun3
The aim of new tests is to check the following:
1- The new dimension clustering idea is fast enough so that we can
apply it even for construction from CHEBFUN3 objects (as opposed to
construction from function handles).

2- Chebfun3 constructor works properly independent of the variable
chosen for separation in Phase 1.

The old dimension clustering implementation (Bebendorf-Kuske idea) is
also removed. The new technique occupies 60% less number of lines of
code.