Draw chebfun3 GUI-based plots in the current figure.
Previously, these functions would create a new figure every time they were
called.  This can be annoying for various reasons, plus, the user _should_ be
able to force the plots to appear in a specific figure if he/she so desires.
In particular, we want to do this ourselves in the plotting test so that we can
put the plots on an invisible figure to prevent them from displaying when the
tests are run.

The solution was to create a wrapper function for openfig() that:

  1. empties the current figure (except for the menu and toolbar; see below)

  2. copies the data from the new figure that openfig() creates into the
     current figure, and

  3. deletes the new figure created by openfig().

The current figure can then be used for everything else.

One subtlety:  it seems that you cannot copy menus and toolbars from one figure
into another in MATLAB.  Therefore, we retain the menu and toolbar on the
current figure instead of copying them from the openfig()-generated one (which,
actually, doesn't have any to begin with for the chebfun3 GUI plots).