Do not make the current figure invisible when opening a chebfun3 GUI plot.
The cosmetic improvement is not really worth the irritation of having
the figure blink in and out of existence briefly while the UI elements
are drawn.  Moreover, it seems that when the figure becomes visible
again at the end of the procedure, its position on the screen changes by
a few pixels, which is most annoying.

This is also more in line with MATLAB plotting, which doesn't do this.