Call newplot() in chebfun3.isosurface().
Built-in isosurface() does not behave like a "normal" plotting function.
In particular, it does not properly clear the figure/axes before drawing
when the plot is not held, so doing, e.g.,

  >> cheb.gallery3('levy')
  >> cheb.gallery3('levy')
  >> cheb.gallery3('levy')

results in three isosurface plots that are "stacked" on top of one
another.  This behavior may make sense for built-in isosurface(), but it
is probably not what we want for chebfun3.isosurface().

This commit fixes this problem by adding a call to newplot() in
chebfun3.isosurface().  newplot() is the function the MATLAB plotting
routines call to handle exactly these same issues; see, e.g., the source
for built-in contour().