Force figures with chebfun3 GUI plots to clear when a new plot is drawn in them.
Prior to this commit, widgets from chebfun3 GUI plots would not get
cleared when a new plot is made.  For instance, doing

  >> cheb.gallery3('runge');
  >> cheb.gallery3('levy');

would result in an isosurface plot for the 'levy' gallery function, but
the sliders for the 'runge' slice plot would still be around.  We fix
this by forcing figures with chebfun3 GUI plots to clear whenever a new
plot is drawn by setting the 'NextPlot' property of the figure to
'replacechildren'.

One caveat is that the 'NextPlot' property must be reset to 'add' at the
beginning of each GUI callback that makes a plot and then must be
returned to 'replacechildren' (or, rather, whatever it was set to, in
case the user has modified it) at the end.  Otherwise, the figure will
be cleared when the callback makes its plot, which is undesirable.  I've
done this by looking at the parent (or, in the case of surf(), the
parent's parent) of the widget whose callback was invoked.  There is
probably a better way to do it.

This commit also simplifies openFigInCurrentFigure() to just do clf()
instead of manually deleting unwanted graphics objects from the current
figure.  Note that we also have to reset the figure's 'NextPlot'
property to 'add' in this function in case the previous plot set it to
'replacechildren'; otherwise, the widgets will be cleared once the GUI
draws its first plot.  Resetting 'NextPlot' should be acceptable here
because we call clf(), so the overall behavior is very similar to
newplot(), which is called by the standard MATLAB plotting functions.
(Indeed, perhaps we should just call newplot() ourselves instead.)