removed singleSignTest from chebfun3/power
CHEBFUN3 cannot represent functions with singularity. Previously when
computing power(f,2) e.g., chebfun3/power used a singleSignTest command
to check if there is a sign change in f and then stopped. As a result,
we could NOT say e.g.
>> cheb.xyz; x.^2
because there is a sign change in x. This is not acceptable as x.^2 is
a perfectly smooth function. Moreover, a user is expected to understand
what she/he is doing. So, we remove the line that calls singleSignTest
in chebfun3/power.