replaced happinessCheck in Chebfun3/constructor with @jaurentz 's suggestion
We need to apply a happinessCheck to the BTD form and we do this one
columns and rows from the first term in the BTD and all the fibres.
Previously, we were adding all such values and called 1D happinessCheck
on the 3 resulting vectors (This was similar to Chebfun2). In this new
version suggested by @jaurentz, instead of adding values, we add
absolute values of coefficients (still from the same first term in BTD).

- 2 more unnecessary roots-related files are also removed.