Use NaN data for the dummy plot in chebfun.plot().
Previously, we just replotted the existing line data and set the plot's
"Visible" property to 'off'.  This doesn't work properly anymore in
R2016b, since MATLAB grays out the legend entries for graphics items
marked invisible.

The fix is easy enough:  we plot NaN data for the dummy plot (but set
the line and point styles appropriately) and leave "Visible" set to
'on'.  MATLAB doesn't draw anything on the plot for NaN data, so we're
good.  A little ugly, but we already use this trick elsewhere for other
things.

Closes #2061.