Properly build q for arbitrary points when nu = 0 in ratinterp().
The problem was that the lines that build q in this case did not
compensate for the R factor from the QR of the Vandermonde matrix.
Without this, we get embarrassing results like this:

  >> [p, q, r] = ratinterp(ones(5, 1), 1, 1, [], chebpts(5));
  >> p(1)/q(1)
  ans =
      -4.472135954999578e-01

With this commit in place, the result is as expected:

  >> [p, q, r] = ratinterp(ones(5, 1), 1, 1, [], chebpts(5));
  >> p(1)/q(1)
  ans =
       9.999999999999997e-01

This error probably went undetected because the code does compensate for
R when nu > 0, and given how often we play with arbitrary points, it's
conceivable that we just never hit the nu = 0 case in all of our
testing.