Fix bracketing bug in stringParser.parSimp() related to division.
We all know that bracketing matters in expressions involving subtraction:
a - b - c and a - (b - c) are not the same thing.  The same is true of
division:  a/b/c = (a/b)/c and a/(b/c) are not the same thing.

stringParser currently handles subtraction correctly but not division.
The parSimp() function strips the brackets from a/(b/c), rendering it as
a/b/c.  This has embarrassing consequences:  if you solve y' = 2/(3/4)
in chebgui, it returns a line with slope 2/3/4 = 0.1667 instead of one
of slope 2/(3/4) = 2.1667.

That example may seem contrived, but I ran across this issue when trying
to solve y' = 1/(2*(y - 1)) with y(0) = 1 and was absolutely stunned
when chebgui gave me the constant solution y = 1 back (instead of
crashing or otherwise failing due to the singularity).  This problem was
a simplified version of another problem shown to me by my colleague
Carlos Borges, who was getting weird answers for it out of chebgui.

This commit fixes this situation by applying the same logic used to
handle subtraction in parSimp() to division:  if we have a division
operator to the left of the brackets we're considering for removal and
find a multiplication or division operator inside (addition and
subtraction already having been handled), we leave the brackets in
place.

This closes #2357.