Create clsap
Classical least squares approximation problem.

A new algorithm for the classical least-squares approximation of N samples by an algebraic polynomial of degree at most n when the number N of the samples is very large. For a fixed n, the complexity of this algorithm in double precision accuracy is O(1). It is faster and more precise than POLYFIT() command.

Includes GRAMPTS() to compute the nodes and weights of Gaussian type quadrature for sums.

See: https://doi.org/10.1137/19M1259936 .